+++ tests/syncrhonizers/test_events_synchronizer.py
@@ -8,7 +8,7 @@ from splitio.api.client import HttpResponse
 from splitio.api import APIException
 from splitio.storage import EventStorage
 from splitio.models.events import Event
-from splitio.synchronizers.event import EventSynchronizer
+from splitio.sync.event import EventSynchronizer
 
 
 class EventsSynchronizerTests(object):
