+++ tests/syncrhonizers/test_impressions_count_synchronizer.py
@@ -8,7 +8,7 @@ from splitio.api.client import HttpResponse
 from splitio.api import APIException
 from splitio.engine.impressions import Manager as ImpressionsManager
 from splitio.engine.impressions import Counter
-from splitio.synchronizers.impression import ImpressionsCountSynchronizer
+from splitio.sync.impression import ImpressionsCountSynchronizer
 from splitio.api.impressions import ImpressionsAPI
